requirements.txt
# Core LangChain packages
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-groq>=0.2.0

# Document processing
pypdf>=4.0.1
pymupdf>=1.24.0

# Embeddings and vector databases
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
chromadb>=0.5.0

# Search and indexing
typesense>=0.19.0

# Web framework and API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0

# Environment and utilities
python-dotenv>=1.0.0

# Build tools (only if needed for development)
setuptools>=75.0.0